requirements.txt
torch==1.13.0
transformers==4.25.1
sentence-transformers==2.2.2
pandas==1.5.2
faiss-cpu==1.7.3
numpy==1.23.5
folium==0.14.0
streamlit==1.16.0
regex == 2.5.123